Edited createTransfer to take in transfer_type_id.
This id determines whether the transfer is pending or not.
Based on this, if we are sending money we check the AccountDao to see if the transfer is possible with senders funds.
Then edit the balances of each account, and create a transfer record.

If we are requesting money, we simply create a pending transfer record, without editing balances.