Moved the validation and sending of funds to TransferDao
createTransfer was originally sending informatino to the accountDao, which was unnecessary. Now, it validates that the sender has enough funds, whether the transfer is a send or request. The funds are then moved via the TransferDao, and instead of resetting the users balances directly, the database is tasked to subtract the transfer amount from the current balance, whatever it may be. This was done because if any server issues happen, and a transfer is delayed the balance could change. So we don't want to calculate the new balance to set it to in Java, rather let the database work out the subtraction so that its always current.